replaced dagger with hilt

+++ app/src/main/java/ke/choxxy/wordsearch/commons/DurationFormatter.java
@@ -8,12 +8,12 @@ import java.util.Locale;
 
 public abstract class DurationFormatter {
 
-    public static String fromInteger(int duration) {
+    public static String fromInteger(long duration) {
         if (duration <= 0)
             return "00:00";
 
-        int secs = duration % 60;
-        int min = duration / 60;
+        long secs = duration % 60;
+        long min = duration / 60;
 
         return String.format(Locale.US, "%02d:%02d", min, secs);
     }
